fix api host model maintain type
Error will occur when get hostgroup info by id:

  sql: Scan error on column index 5: converting driver.Value type int64 ("1497951907") to a uint16: value out of range

+++ modules/api/app/model/falcon_portal/host.go
@@ -26,8 +26,8 @@ type Host struct {
 	Ip            string `json:"ip" gorm:"column:ip"`
 	AgentVersion  string `json:"agent_version"  gorm:"column:agent_version"`
 	PluginVersion string `json:"plugin_version"  gorm:"column:plugin_version"`
-	MaintainBegin uint16 `json:"maintain_begin"  gorm:"column:maintain_begin"`
-	MaintainEnd   uint16 `json:"maintain_end"  gorm:"column:maintain_end"`
+	MaintainBegin int64  `json:"maintain_begin"  gorm:"column:maintain_begin"`
+	MaintainEnd   int64  `json:"maintain_end"  gorm:"column:maintain_end"`
 }
 
 func (this Host) TableName() string {
